refactor(commons): convert FileModule module into functions
